+++ lib/commands/devices/commands/add_command.dart
@@ -2,14 +2,12 @@
 
 import 'dart:async';
 
-import 'package:flutter_tools/src/base/io.dart';
-import 'package:flutter_tools/src/custom_devices/custom_device_config.dart';
-import 'package:snapp_cli/commands/base_command.dart';
-import 'package:snapp_cli/configs/predefined_devices.dart';
-import 'package:snapp_cli/host_runner/host_runner_platform.dart';
+import 'package:snapp_cli/service/custom_device_builder/custom_device_builder.dart';
 import 'package:snapp_cli/service/remote_controller_service.dart';
+import 'package:snapp_cli/service/setup_device/device_setup.dart';
+import 'package:snapp_cli/service/setup_device/src/custom_embedder_provider.dart';
+import 'package:snapp_cli/service/setup_device/src/install_dependency_provider.dart';
 import 'package:snapp_cli/service/ssh_service.dart';
-import 'package:snapp_cli/utils/common.dart';
 
 /// Add a new raspberry device to the Flutter SDK custom devices
 class AddCommand extends BaseSnappCommand {
@@ -27,216 +25,62 @@ class AddCommand extends BaseSnappCommand {
   final name = 'add';
 
   @override
-  final description = 'add a new device to custom devices';
+  final description = 'add a new device to custom devices manually.';
 
   @override
   Future<int> run() async {
     logger.spaces();
 
-    final addCommandOptions = [
-      'Express (recommended)',
-      'Custom',
-    ];
+    logger.warn('''
+Add command is a way to add a new device to custom devices manually.
 
-    final commandIndex = interaction.selectIndex(
-      'Please select the type of device you want to add.',
-      options: addCommandOptions,
-    );
+We recommend you to use the bootstrap command to setup a device from scratch.
 
-    if (commandIndex == 0) {
-      return _addPredefinedDevice();
-    }
-
-    return _addCustomDevice();
-  }
-
-  Future<int> _addPredefinedDevice() async {
-    logger.spaces();
-
-    final deviceKey = interaction.select(
-      'Select your device',
-      options: predefinedDevices.keys.toList(),
-    );
-
-    var predefinedDeviceConfig = predefinedDevices[deviceKey];
-
-    if (predefinedDeviceConfig == null) {
-      throwToolExit(
-          'Something went wrong while trying to add predefined $deviceKey device.');
-    }
-
-    /// check if the device id already exists in the config file
-    /// update the id if it
-    if (_isDuplicatedDeviceId(predefinedDeviceConfig.id)) {
-      predefinedDeviceConfig = predefinedDeviceConfig.copyWith(
-        id: _suggestIdForDuplicatedDeviceId(predefinedDeviceConfig.id),
-      );
-    }
-
-    return _addCustomDevice(predefinedConfig: predefinedDeviceConfig);
-  }
-
-  Future<int> _addCustomDevice({CustomDeviceConfig? predefinedConfig}) async {
-    logger.spaces();
-
-    // get remote device id and label from the user
-    final id = predefinedConfig?.id.isNotEmpty == true
-        ? predefinedConfig!.id
-        : interaction.readDeviceId(customDevicesConfig);
-
-    final label = predefinedConfig?.label.isNotEmpty == true
-        ? predefinedConfig!.label
-        : interaction.readDeviceLabel();
-
-    // get remote device ip and username from the user
-    logger.spaces();
-
-    logger.info('to add a new device, we need an IP address and a username.');
-
-    final targetIp = interaction.readDeviceIp(
-        description:
-            'Please enter the IP-address of the device. (example: 192.168.1.101)');
-
-    final username = interaction.readDeviceUsername(
-      description:
-          'Please enter the username used for ssh-ing into the remote device. (example: pi)',
-    );
-
-    final bool ipv6 = targetIp.isIpv6;
-
-    final InternetAddress loopbackIp =
-        ipv6 ? InternetAddress.loopbackIPv6 : InternetAddress.loopbackIPv4;
-
-    // SSH expects IPv6 addresses to use the bracket syntax like URIs do too,
-    // but the IPv6 the user enters is a raw IPv6 address, so we need to wrap it.
-    final String sshTarget = targetIp.sshTarget(username);
-
-    final String formattedLoopbackIp =
-        ipv6 ? '[${loopbackIp.address}]' : loopbackIp.address;
+You need to provide specific PATH to the custom embedder you want to use.
+    ''');
 
     logger.spaces();
 
-    logger.info(
-      '''
-We need the exact path of your flutter command line tools on the remote device. 
-We will use this path to run flutter commands on the remote device like "flutter build linux --debug". 
-''',
-    );
+    DeviceSetup deviceSetup = DeviceSetup(
+      steps: [
+        /// Receives information about the target device like id, name and type.
+        /// Example: Raspberry Pi 4b
+        DeviceTypeProvider(customDevicesConfig: customDevicesConfig),
 
-    logger.spaces();
+        /// Receives connection information about the target device like ip, port, username.
+        DeviceHostProvider(),
 
-    String remoteRunnerCommand = await interaction.readFlutterManualPath();
+        /// Checks what kind of embedder user wants to use.
+        /// Example: Flutter, Flutter-pi ...
+        CustomEmbedderProvider(),
 
-    /// path to the icu data file on the host machine
-    final hostIcuDataPath = flutterSdkManager.icuDataPath;
+        /// Installs dependencies required to run the app on the remote device.
+        /// Regarding to the embedder type
+        /// for example for Flutter-pi, it will install flutterpi_tool global package
+        InstallDependencyProvider(
+          remoteControllerService: remoteControllerService,
+          flutterSdkManager: flutterSdkManager,
+        ),
 
-    /// path to the build artifacts on the remote machine
-    const hostBuildClonePath = 'snapp_embedded';
-
-    /// path to the icu data file on the remote machine
-    const hostIcuDataClone = '$hostBuildClonePath/engine';
-
-    CustomDeviceConfig config = CustomDeviceConfig(
-      id: id,
-      label: label,
-      sdkNameAndVersion: label,
-      enabled: true,
-
-      // host-platform specific, filled out later
-      pingCommand:
-          hostPlatform.pingCommand(ipv6: ipv6, pingTarget: targetIp.address),
-      pingSuccessRegex: hostPlatform.pingSuccessRegex,
-      postBuildCommand: const <String>[],
-
-      // just install to /tmp/${appName} by default
-      // returns the command runner for the current platform
-      // for example:
-      // on windows it returns "powershell -c"
-      // on linux and macOS it returns "bash -c"
-      installCommand: hostPlatform.commandRunner(
-        <String>[
-          // create the necessary directories in the remote machine
-          hostPlatform
-              .sshCommand(
-                ipv6: ipv6,
-                sshTarget: sshTarget,
-                command: 'mkdir -p /tmp/\${appName}/$hostIcuDataClone',
-              )
-              .asString,
-
-          // copy the current project files from host to the remote
-          hostPlatform
-              .scpCommand(
-                ipv6: ipv6,
-                source: '${hostPlatform.currentSourcePath}*',
-                dest: '$sshTarget:/tmp/\${appName}',
-              )
-              .asString,
-
-          // copy the build artifacts from host to the remote
-          hostPlatform
-              .scpCommand(
-                ipv6: ipv6,
-                source: r'${localPath}',
-                dest: '$sshTarget:/tmp/\${appName}/$hostBuildClonePath',
-              )
-              .asString,
-
-          // copy the icu data file from host to the remote
-          hostPlatform
-              .scpCommand(
-                ipv6: ipv6,
-                source: hostIcuDataPath,
-                dest: '$sshTarget:/tmp/\${appName}/$hostIcuDataClone',
-                lastCommand: true,
-              )
-              .asString,
-        ],
-      ),
-      // just uninstall app by removing the /tmp/${appName} directory on the remote
-      uninstallCommand: hostPlatform.sshCommand(
-        ipv6: ipv6,
-        sshTarget: sshTarget,
-        command: r'rm -rf "/tmp/${appName}"',
-        lastCommand: true,
-      ),
-
-      // run the app on the remote
-      runDebugCommand: hostPlatform.sshMultiCommand(
-        ipv6: ipv6,
-        sshTarget: sshTarget,
-        commands: <String>[
-          'cd /tmp/\${appName} ;',
-          '$remoteRunnerCommand build linux --debug ;',
-          // remove remote build artifacts
-          'rm -rf "/tmp/\${appName}/build/flutter_assets/*" ;',
-          'rm -rf "/tmp/\${appName}/build/linux/arm64/debug/bundle/data/flutter_assets/*" ;',
-          'rm -rf "/tmp/\${appName}/build/linux/arm64/debug/bundle/data/icudtl.dat" ;',
-          // and replace them by host build artifacts
-          'cp /tmp/\${appName}/$hostBuildClonePath/flutter_assets/*  /tmp/\${appName}/build/flutter_assets ;',
-          'cp /tmp/\${appName}/$hostBuildClonePath/flutter_assets/*  /tmp/\${appName}/build/linux/arm64/debug/bundle/data/flutter_assets ;',
-          'cp /tmp/\${appName}/$hostIcuDataClone/icudtl.dat  /tmp/\${appName}/build/linux/arm64/debug/bundle/data ;',
-          // finally run the app
-          r'DISPLAY=:0 /tmp/\${appName}/build/linux/arm64/debug/bundle/\${appName} ;'
-        ],
-      ),
-      forwardPortCommand: <String>[
-        'ssh',
-        '-o',
-        'BatchMode=yes',
-        '-o',
-        'ExitOnForwardFailure=yes',
-        if (ipv6) '-6',
-        '-L',
-        '$formattedLoopbackIp:\${hostPort}:$formattedLoopbackIp:\${devicePort}',
-        sshTarget,
-        "echo 'Port forwarding success'; read",
+        /// installs custom embedder chosen by the user.
+        ManualAppExecuterProvider(),
       ],
-      forwardPortSuccessRegex: RegExp('Port forwarding success'),
-      screenshotCommand: null,
     );
 
-    customDevicesConfig.add(config);
+    final deviceContext = (await deviceSetup.setup()).copyWith(
+      // We need to set this to true because we are not setting up passwordless ssh connection in this command.
+      remoteHasSshConnection: true,
+    );
+
+    final deviceBuilder = CustomDeviceBuilder.create(
+      embedder: deviceContext.embedder!,
+      flutterSdkManager: flutterSdkManager,
+      hostPlatform: hostPlatform,
+    );
+
+    final newDeviceConfig = await deviceBuilder.buildDevice(deviceContext);
+
+    customDevicesConfig.add(newDeviceConfig);
 
     logger.spaces();
 
@@ -247,24 +91,5 @@ We will use this path to run flutter commands on the remote device like "flutter
     logger.spaces();
 
     return 0;
-  }
-
-  bool _isDuplicatedDeviceId(String s) {
-    return customDevicesConfig.devices.any((element) => element.id == s);
-  }
-
-  /// returns new device id by adding a number to the end of it
-  ///
-  /// for example: if the id is "pi-4" and this id already exists
-  /// then we update the id to "pi-4-1" and check again
-  /// if the new id already exists, then we update it to "pi-4-2" and so on
-  String _suggestIdForDuplicatedDeviceId(String s) {
-    int i = 1;
-
-    while (_isDuplicatedDeviceId('$s-$i')) {
-      i++;
-    }
-
-    return '$s-$i';
   }
 }